Added basic output changing script
test.py currently reads the file 100_ind and 10d20c.data and processes 10d20c.data using the algorithms from 100_ind in order to create a new data file with a smaller number of features. Because the algorithms in 100_ind were in the same format as python functions, test.py creates a new file, action.py, and adds the algorithms in 100_ind as well as the implementations of the functions in the algorithms in 100_ind. It then runs action.py while continuing to use 10d20c.data to create the file result.txt